+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - chore: update layout and register in G6;
 - fix: performance problem in create-edge with polyline;
 - fix: performance for polyline;
-- fix: throttle updating the polyline edges in drag-node behavior;
+- fix: debounce updating the polyline edges in drag-node behavior;
 - fix: toolbar redo undo max clone in drag-node behavior;
 
 #### 4.1.12